chore: clean pd mla pr surface
