+++ python/pegaflow/pegaflow.pyi
@@ -310,15 +310,6 @@ class PyLoadState:
         """
         ...
 
-class PdRdmaTestBuffer:
-    def __init__(self, *, size: int, cuda_device: int = 0) -> None: ...
-    def ptr(self) -> int: ...
-    def size(self) -> int: ...
-    def fill(self, value: int) -> None: ...
-    def to_bytes(self) -> bytes: ...
-    def to_bytes_range(self, offset: int, len: int) -> bytes: ...
-    def write_bytes(self, offset: int, data: bytes) -> None: ...
-
 class PdRdmaEngine:
     def __init__(
         self,
