+++ python/tests/test_pd_connector_flow.py
@@ -1,748 +1,7 @@
 from __future__ import annotations
 
-import queue
-import sys
-import threading
-import time
-import types
-from types import SimpleNamespace
-from unittest.mock import MagicMock
-
-import pytest
-
-from .unit_stubs import install_connector_unit_stubs
-
-install_connector_unit_stubs()
-
-from vllm.distributed.kv_transfer.kv_connector.v1.base import (  # noqa: E402
-    KVConnectorRole,
-)
-
-native = types.ModuleType("pegaflow.pegaflow")
-native.EngineRpcClient = MagicMock
-native.PegaFlowError = type("PegaFlowError", (Exception,), {})
-native.PegaflowInternal = type("PegaflowInternal", (Exception,), {})
-native.PyLoadState = object
-native.QueryLoading = object
-native.QueryReady = object
-native.__version__ = "test"
-sys.modules["pegaflow.pegaflow"] = native
-
-import pegaflow.pd_connector.decode_worker as decode_worker_mod  # noqa: E402
-import pegaflow.pd_connector.prefill as prefill_mod  # noqa: E402
-import pegaflow.pd_connector.prefill_worker as prefill_worker_mod  # noqa: E402
-import pegaflow.pd_connector.worker as worker_mod  # noqa: E402
-from pegaflow.pd_connector import PdConnector  # noqa: E402
-from pegaflow.pd_connector.kv_params import parse_consumer  # noqa: E402
-from pegaflow.pd_connector.layout import (  # noqa: E402
-    BlockRegionSlice,
-    FlashAttnHndLayout,
-    LayerBlockSlices,
-    unique_blocks_from_slot_mapping,
-)
-from pegaflow.pd_connector.metadata import (  # noqa: E402
-    LayerRemoteLayout,
-    PdConnectorMetadata,
-    PdHandshake,
-    PushReqMeta,
-    TransferRegionLayout,
-    WaitReqMeta,
-    handshake_from_dict,
-    handshake_to_compact_dict,
-    handshake_to_dict,
-)
-from pegaflow.pd_connector.prefill import (  # noqa: E402
-    AsyncPrefillSender,
-    PrefillHttpTask,
-)
-from pegaflow.pd_connector.proxy import (  # noqa: E402
-    ProxyConfig,
-    build_pd_proxy_request,
-)
-from pegaflow.pd_connector.rdma import (  # noqa: E402
-    MockRdmaPort,
-    RealRdmaPort,
-    _layer_blocks_to_native,
-)
-from pegaflow.pd_connector.scheduler import PdSchedulerConnector  # noqa: E402
-from pegaflow.pd_connector.worker import PdWorkerConnector  # noqa: E402
-
-
-def teardown_module() -> None:
-    sys.modules.pop("pegaflow.pegaflow", None)
-
-
-class FakeTensor:
-    def __init__(
-        self,
-        shape: tuple[int, ...],
-        stride: tuple[int, ...],
-        ptr: int = 0x1000,
-        element_size: int = 2,
-        device_index: int | None = None,
-    ) -> None:
-        self.shape = shape
-        self._stride = stride
-        self._ptr = ptr
-        self._element_size = element_size
-        self.device = SimpleNamespace(index=device_index) if device_index is not None else None
-
-    def stride(self) -> tuple[int, ...]:
-        return self._stride
-
-    def data_ptr(self) -> int:
-        return self._ptr
-
-    def element_size(self) -> int:
-        return self._element_size
-
-
-class FakeSlotMapping(list[int]):
-    def __init__(self, values: list[int]) -> None:
-        super().__init__(values)
-        self.cpu_calls = 0
-
-    def detach(self):
-        return self
-
-    def cpu(self):
-        self.cpu_calls += 1
-        return self
-
-    def tolist(self):
-        return list(self)
-
-
-class FakePrefillSender:
-    def __init__(self) -> None:
-        self.tasks = []
-
-    def submit(self, task) -> None:
-        self.tasks.append(task)
-
-
-class FakeNativeRdmaEngine:
-    def __init__(self) -> None:
-        self.local_layers = []
-        self.remote_regs = []
-        self.pushed_layers = []
-        self.done_reqs = []
-        self.waited_reqs = []
-        self.polled_reqs = []
-        self.marked_reqs = []
-        self.finished_sending = ["sent-1"]
-        self.finished_recving = ["recv-1"]
-
-    def register_local_layers(self, layers):
-        self.local_layers.append(layers)
-        registered = []
-        for layer in layers:
-            assert "regions" in layer
-            assert "k_block_addrs" not in layer
-            assert "v_block_addrs" not in layer
-            ptr = min(region["base_addr"] for region in layer["regions"])
-            registered.append(
-                {
-                    **layer,
-                    "mr_desc": {
-                        "ptr": ptr,
-                        "addr_rkey_list": [["10.0.0.1:1", 17]],
-                    },
-                }
-            )
-        return registered
-
-    def register_remote(self, req_id, handshake):
-        assert handshake is not None
-        assert handshake["request_id"]
-        assert handshake.get("imm_id") is None or isinstance(handshake["imm_id"], int)
-        for layer in handshake["layers"]:
-            assert layer["mr_desc"]["addr_rkey_list"]
-            assert isinstance(layer["mr_desc"]["addr_rkey_list"][0], tuple)
-            assert layer["block_ids"]
-            assert layer["regions"]
-            assert [region["region_idx"] for region in layer["regions"]] == list(
-                range(len(layer["regions"]))
-            )
-            assert "k_block_addrs" not in layer
-            assert "v_block_addrs" not in layer
-        self.remote_regs.append((req_id, handshake))
-
-    def push_layer(self, req_id, layer_idx, blocks):
-        for block in blocks:
-            assert set(block) == {"regions"}
-            assert [region["region_idx"] for region in block["regions"]] == list(
-                range(len(block["regions"]))
-            )
-        self.pushed_layers.append((req_id, layer_idx, blocks))
-
-    def push_done(self, req_id):
-        self.done_reqs.append(req_id)
-
-    def wait_done(self, req_id):
-        self.waited_reqs.append(req_id)
-
-    def poll_done(self, req_id):
-        self.polled_reqs.append(req_id)
-        return req_id in self.finished_recving
-
-    def mark_done(self, req_id):
-        self.marked_reqs.append(req_id)
-
-    def pop_finished_sending(self):
-        finished = self.finished_sending
-        self.finished_sending = []
-        return finished
-
-    def pop_finished_recving(self):
-        finished = self.finished_recving
-        self.finished_recving = []
-        return finished
-
-
-class FakeNativeRdmaEngineCtor(FakeNativeRdmaEngine):
-    last_kwargs = None
-
-    def __init__(self, **kwargs) -> None:
-        super().__init__()
-        type(self).last_kwargs = kwargs
-
-    def num_domains(self):
-        return 1
-
-    def num_groups(self):
-        return 1
-
-    def aggregated_link_speed(self):
-        return 400_000_000_000
-
-
-def drain_pd_pushes(worker: PdWorkerConnector) -> None:
-    worker._push_sender.wait_all()
-    worker._push_finalizer.wait_all()
-
-
-def pushed_layers_by_idx(
-    rdma: MockRdmaPort,
-    req_id: str,
-) -> dict[int, list[LayerBlockSlices]]:
-    return dict(rdma.pushed_layers[req_id])
-
-
-DUMMY_HANDSHAKE = PdHandshake(
-    request_id="",
-    engine_id="",
-    tp_rank=0,
-    tp_size=1,
-    block_size=16,
-    layers=(),
-)
-
-
-def hnd_remote_layer(
-    *,
-    layer_name: str = "layer.0",
-    layer_idx: int = 0,
-    block_ids: tuple[int, ...] = (0,),
-    k_base: int = 0x1000,
-    v_base: int = 0x8000,
-    block_len: int = 0x400,
-) -> LayerRemoteLayout:
-    return LayerRemoteLayout(
-        layer_name=layer_name,
-        layer_idx=layer_idx,
-        block_ids=block_ids,
-        regions=(
-            TransferRegionLayout(region_idx=0, base_addr=k_base, block_len=block_len),
-            TransferRegionLayout(region_idx=1, base_addr=v_base, block_len=block_len),
-        ),
-    )
-
-
-def fake_mla_config(
-    *,
-    tp_rank: int = 0,
-    tp_size: int = 1,
-    block_size: int = 64,
-) -> SimpleNamespace:
-    return SimpleNamespace(
-        kv_transfer_config=SimpleNamespace(engine_id="pd"),
-        model_config=SimpleNamespace(use_mla=True),
-        cache_config=SimpleNamespace(block_size=block_size),
-        parallel_config=SimpleNamespace(
-            tensor_parallel_rank=tp_rank,
-            tensor_parallel_size=tp_size,
-            decode_context_parallel_size=1,
-            prefill_context_parallel_size=1,
-        ),
-    )
-
-
-def fake_kv_cache_config(
-    *,
-    num_blocks: int,
-    specs: dict[str, SimpleNamespace],
-) -> SimpleNamespace:
-    return SimpleNamespace(
-        num_blocks=num_blocks,
-        kv_cache_groups=[
-            SimpleNamespace(
-                layer_names=tuple(specs),
-                kv_cache_spec=SimpleNamespace(kv_cache_specs=specs),
-            )
-        ],
-    )
-
-
-def test_flash_attn_hnd_layout_offsets() -> None:
-    # Logical shape [2, num_blocks, block_size, num_kv_heads, head_size].
-    # HND physical order [2, num_blocks, num_kv_heads, block_size, head_size].
-    tensor = FakeTensor(
-        shape=(2, 8, 16, 4, 32),
-        stride=(8 * 4 * 16 * 32, 4 * 16 * 32, 32, 16 * 32, 1),
-    )
-
-    layout = FlashAttnHndLayout.from_tensor("layer.0", tensor)
-
-    assert layout.block_bytes == 16 * 4 * 32 * 2
-    assert layout.block_offset_bytes(0, 3) == 3 * 4 * 16 * 32 * 2
-    assert layout.block_offset_bytes(1, 3) == (8 * 4 * 16 * 32 + 3 * 4 * 16 * 32) * 2
-
-
-def test_pd_worker_registers_mla_and_indexer_layouts_from_layer_specs() -> None:
-    main_tensor = FakeTensor(
-        shape=(8, 64, 656),
-        stride=(64 * 656, 656, 1),
-        ptr=0x1000,
-        element_size=1,
-    )
-    indexer_tensor = FakeTensor(
-        shape=(8, 64, 128),
-        stride=(64 * 128, 128, 1),
-        ptr=0x200000,
-        element_size=1,
-    )
-    kv_cache_config = fake_kv_cache_config(
-        num_blocks=8,
-        specs={
-            "layer.0": SimpleNamespace(block_size=64, page_size_bytes=64 * 656),
-            "indexer.0": SimpleNamespace(block_size=64, page_size_bytes=64 * 128),
-        },
-    )
-    worker = PdWorkerConnector(
-        fake_mla_config(),
-        kv_cache_config=kv_cache_config,
-        rdma=MockRdmaPort(),
-    )
-
-    worker.register_kv_caches({"layer.0": main_tensor, "indexer.0": indexer_tensor})
-
-    assert worker.layouts["layer.0"].block_size == 64
-    assert worker.layouts["layer.0"].remote_layout(0).regions == (
-        TransferRegionLayout(region_idx=0, base_addr=0x1000, block_len=64 * 656),
-    )
-    assert worker.layouts["indexer.0"].remote_layout(1).regions == (
-        TransferRegionLayout(region_idx=0, base_addr=0x200000, block_len=64 * 128),
-    )
-    assert (
-        worker.rdma.local_layers[0].regions[0].block_len
-        != worker.rdma.local_layers[1].regions[0].block_len
-    )
-
-
-def test_pd_connector_mla_returns_default_layout_and_allows_128_block_size() -> None:
-    assert PdConnector.get_required_kvcache_layout(fake_mla_config(block_size=64)) is None
-    PdConnector(fake_mla_config(block_size=128), KVConnectorRole.WORKER)
-
-
-def test_pd_worker_rejects_mla_physical_logical_block_split() -> None:
-    tensor = FakeTensor(
-        shape=(8, 32, 128),
-        stride=(32 * 128, 128, 1),
-        ptr=0x1000,
-        element_size=1,
-    )
-    kv_cache_config = fake_kv_cache_config(
-        num_blocks=8,
-        specs={
-            "layer.0": SimpleNamespace(block_size=64, page_size_bytes=32 * 128),
-        },
-    )
-    worker = PdWorkerConnector(
-        fake_mla_config(block_size=64),
-        kv_cache_config=kv_cache_config,
-        rdma=MockRdmaPort(),
-    )
-
-    with pytest.raises(AssertionError, match="physical/logical block split"):
-        worker.register_kv_caches({"layer.0": tensor})
-
-
-def test_p_worker_maps_mla_prefill_tp_greater_than_decode_tp() -> None:
-    handshakes = tuple(
-        PdHandshake(
-            request_id=f"decode-r{rank}",
-            engine_id="decode",
-            tp_rank=rank,
-            tp_size=4,
-            block_size=64,
-            layers=(),
-        )
-        for rank in range(4)
-    )
-    worker = PdWorkerConnector(
-        fake_mla_config(tp_rank=2, tp_size=8),
-        rdma=MockRdmaPort(),
-    )
-
-    worker.start_load_kv(
-        PdConnectorMetadata(
-            reqs_to_push={
-                "prefill-r2": PushReqMeta(
-                    local_block_ids=([1],),
-                    target_request_id="decode",
-                    handshakes=handshakes,
-                )
-            }
-        ),
-        None,
-    )
-
-    assert worker.rdma.remote_handshakes["prefill-r2"] is handshakes[1]
-
-
-def test_p_worker_skips_non_representative_mla_prefill_rank() -> None:
-    handshakes = tuple(
-        PdHandshake(
-            request_id=f"decode-r{rank}",
-            engine_id="decode",
-            tp_rank=rank,
-            tp_size=4,
-            block_size=64,
-            layers=(),
-        )
-        for rank in range(4)
-    )
-    worker = PdWorkerConnector(
-        fake_mla_config(tp_rank=3, tp_size=8),
-        rdma=MockRdmaPort(),
-    )
-
-    worker.start_load_kv(
-        PdConnectorMetadata(
-            reqs_to_push={
-                "prefill-r3": PushReqMeta(
-                    local_block_ids=([1],),
-                    target_request_id="decode",
-                    handshakes=handshakes,
-                )
-            }
-        ),
-        None,
-    )
-
-    assert "prefill-r3" not in worker.rdma.remote_handshakes
-    assert worker.get_finished({"prefill-r3"}) == ({"prefill-r3"}, None)
-
-
-def test_real_rdma_port_preserves_native_contract_for_pd_push() -> None:
-    native_engine = FakeNativeRdmaEngine()
-    rdma = RealRdmaPort(native_engine)
-    layer = hnd_remote_layer(
-        block_ids=(0, 1),
-        k_base=0x1000,
-        v_base=0x9000,
-        block_len=4096,
-    )
-
-    registered = rdma.register_local_layers((layer,))
-
-    assert registered[0].mr_desc == {
-        "ptr": 0x1000,
-        "addr_rkey_list": [["10.0.0.1:1", 17]],
-    }
-    assert registered[0].block_ids == (0, 1)
-    assert registered[0].regions == layer.regions
-
-    handshake = PdHandshake(
-        request_id="req-1",
-        engine_id="decode",
-        tp_rank=0,
-        tp_size=1,
-        block_size=16,
-        layers=registered,
-    )
-    rdma.open_request("req-1", handshake)
-
-    assert native_engine.remote_regs[0][0] == "req-1"
-    assert native_engine.remote_regs[0][1]["layers"][0]["mr_desc"]["ptr"] == 0x1000
-    assert native_engine.remote_regs[0][1]["layers"][0]["block_ids"] == [0, 1]
-    assert native_engine.remote_regs[0][1]["layers"][0]["regions"] == [
-        {"region_idx": 0, "base_addr": 0x1000, "block_len": 4096},
-        {"region_idx": 1, "base_addr": 0x9000, "block_len": 4096},
-    ]
-
-    rdma.push_layer(
-        "req-1",
-        0,
-        [
-            FlashAttnHndLayout(
-                "layer.0", (2, 8, 16, 4, 32), (16384, 2048, 32, 512, 1), 2, 0x1000
-            ).block_slices(1)
-        ],
-    )
-    rdma.push_done("req-1")
-    rdma.wait_done("req-1")
-    rdma.mark_done("req-1")
-
-    _, _, pushed_blocks = native_engine.pushed_layers[0]
-    assert pushed_blocks == [
-        {
-            "regions": [
-                {
-                    "region_idx": 0,
-                    "block_id": 1,
-                    "src_offset_bytes": 2048 * 2,
-                    "bytes": 4096,
-                },
-                {
-                    "region_idx": 1,
-                    "block_id": 1,
-                    "src_offset_bytes": (16384 + 2048) * 2,
-                    "bytes": 4096,
-                },
-            ],
-        }
-    ]
-    assert native_engine.done_reqs == ["req-1"]
-    assert native_engine.waited_reqs == ["req-1"]
-    assert native_engine.marked_reqs == ["req-1"]
-    assert rdma.pop_finished_sending() == {"sent-1"}
-    assert rdma.pop_finished_sending() == set()
-    assert rdma.pop_finished_recving() == {"recv-1"}
-    assert rdma.pop_finished_recving() == set()
-
-
-def test_real_rdma_port_rejects_native_layout_without_block_mapping() -> None:
-    native_engine = FakeNativeRdmaEngine()
-
-    def broken_register_local_layers(layers):
-        registered = [{**layers[0], "mr_desc": {"ptr": 0x1000, "addr_rkey_list": []}}]
-        registered[0].pop("block_ids")
-        return registered
-
-    native_engine.register_local_layers = broken_register_local_layers
-    rdma = RealRdmaPort(native_engine)
-    layer = hnd_remote_layer(block_ids=(0,), block_len=1024)
-
-    with pytest.raises(KeyError, match="block_ids"):
-        rdma.register_local_layers((layer,))
-
-
-def test_pd_handshake_serializes_regions_layout() -> None:
-    layer = hnd_remote_layer(
-        block_ids=(8, 9, 10),
-        k_base=0x10_000,
-        v_base=0x20_000,
-        block_len=0x400,
-    )
-    handshake = PdHandshake(
-        request_id="req-1",
-        engine_id="decode",
-        tp_rank=0,
-        tp_size=1,
-        block_size=16,
-        layers=(layer,),
-    )
-
-    data = handshake_to_dict(handshake)
-    assert data is not None
-    layer_data = data["layers"][0]
-    assert layer_data["block_ids"] == [8, 9, 10]
-    assert layer_data["regions"] == [
-        {"region_idx": 0, "base_addr": 0x10_000, "block_len": 0x400},
-        {"region_idx": 1, "base_addr": 0x20_000, "block_len": 0x400},
-    ]
-    assert "k_block_addrs" not in layer_data
-    assert "v_block_addrs" not in layer_data
-    assert "linear" not in layer_data
-
-    restored = handshake_from_dict(data)
-    assert restored is not None
-    assert restored.layers[0].block_ids == layer.block_ids
-    assert restored.layers[0].regions == layer.regions
-    assert restored.layers[0].region_block_addrs(0) == (
-        0x10_000 + 8 * 0x400,
-        0x10_000 + 9 * 0x400,
-        0x10_000 + 10 * 0x400,
-    )
-
-
-def test_pd_handshake_compact_serializes_shared_block_ids_once() -> None:
-    layers = (
-        hnd_remote_layer(layer_name="layer.0", layer_idx=0, block_ids=(8, 9, 10)),
-        hnd_remote_layer(layer_name="layer.1", layer_idx=1, block_ids=(8, 9, 10)),
-    )
-    handshake = PdHandshake(
-        request_id="req-1",
-        engine_id="decode",
-        tp_rank=0,
-        tp_size=1,
-        block_size=16,
-        layers=layers,
-    )
-
-    data = handshake_to_compact_dict(handshake)
-
-    assert data["block_ids"] == [8, 9, 10]
-    assert "block_ids" not in data["layers"][0]
-    assert "block_ids" not in data["layers"][1]
-    restored = handshake_from_dict(data)
-    assert restored is not None
-    assert restored.layers[0].block_ids == (8, 9, 10)
-    assert restored.layers[1].block_ids == (8, 9, 10)
-
-
-def test_pd_worker_builds_native_rdma_by_default_when_extension_exists(monkeypatch) -> None:
-    monkeypatch.setattr(native, "PdRdmaEngine", FakeNativeRdmaEngineCtor, raising=False)
-    tensor = FakeTensor(
-        shape=(2, 8, 16, 4, 32),
-        stride=(8 * 4 * 16 * 32, 4 * 16 * 32, 32, 16 * 32, 1),
-        device_index=2,
-    )
-    config = SimpleNamespace(
-        kv_transfer_config=SimpleNamespace(
-            engine_id="decode",
-            get_from_extra_config=lambda key, default: {
-                "pegaflow.pd.rdma.rank_map": {"0": {"nic": "mlx5_2", "worker_cpu": 64}},
-            }.get(key, default),
-        ),
-        parallel_config=SimpleNamespace(tensor_parallel_rank=0),
-    )
-
-    worker = PdWorkerConnector(config)
-    worker.register_kv_caches({"layer.0": tensor})
-
-    assert isinstance(worker.rdma, RealRdmaPort)
-    assert FakeNativeRdmaEngineCtor.last_kwargs == {
-        "cuda_device": 2,
-        "numa_node": None,
-        "domains": ["mlx5_2"],
-        "device": "cuda",
-        "pin_worker_cpu": 64,
-    }
-
-
-def test_pd_worker_uses_runtime_tp_rank_for_rdma_rank_map(monkeypatch) -> None:
-    monkeypatch.setattr(native, "PdRdmaEngine", FakeNativeRdmaEngineCtor, raising=False)
-    monkeypatch.setattr(worker_mod, "get_tensor_model_parallel_rank", lambda: 2)
-    monkeypatch.setattr(worker_mod, "get_tensor_model_parallel_world_size", lambda: 8)
-    tensor = FakeTensor(
-        shape=(2, 8, 16, 4, 32),
-        stride=(8 * 4 * 16 * 32, 4 * 16 * 32, 32, 16 * 32, 1),
-        device_index=2,
-    )
-    config = SimpleNamespace(
-        kv_transfer_config=SimpleNamespace(
-            engine_id="decode",
-            get_from_extra_config=lambda key, default: {
-                "pegaflow.pd.rdma.rank_map": {
-                    "0": {"nic": "mlx5_1", "worker_cpu": 16},
-                    "2": {"nic": "mlx5_2", "worker_cpu": 60},
-                },
-            }.get(key, default),
-        ),
-        parallel_config=SimpleNamespace(tensor_parallel_rank=0, tensor_parallel_size=8),
-    )
-
-    worker = PdWorkerConnector(config)
-    worker.register_kv_caches({"layer.0": tensor})
-
-    assert FakeNativeRdmaEngineCtor.last_kwargs["domains"] == ["mlx5_2"]
-    assert FakeNativeRdmaEngineCtor.last_kwargs["pin_worker_cpu"] == 60
-
-
-def test_pd_worker_rejects_legacy_global_rdma_config(monkeypatch) -> None:
-    monkeypatch.setattr(native, "PdRdmaEngine", FakeNativeRdmaEngineCtor, raising=False)
-    tensor = FakeTensor(
-        shape=(2, 8, 16, 4, 32),
-        stride=(8 * 4 * 16 * 32, 4 * 16 * 32, 32, 16 * 32, 1),
-        device_index=2,
-    )
-    config = SimpleNamespace(
-        kv_transfer_config=SimpleNamespace(
-            engine_id="decode",
-            get_from_extra_config=lambda key, default: {
-                "pegaflow.pd.rdma.domains": ["mlx5_2"],
-                "pegaflow.pd.rdma.rank_map": {"0": {"nic": "mlx5_2", "worker_cpu": 64}},
-            }.get(key, default),
-        ),
-        parallel_config=SimpleNamespace(tensor_parallel_rank=0),
-    )
-
-    worker = PdWorkerConnector(config)
-    with pytest.raises(RuntimeError, match="legacy keys"):
-        worker.register_kv_caches({"layer.0": tensor})
-
-
-def test_rdma_native_blocks_coalesce_contiguous_ranges() -> None:
-    blocks = [
-        LayerBlockSlices(
-            regions=(
-                BlockRegionSlice(block_id=10, src_offset_bytes=0x1000, bytes=0x400),
-                BlockRegionSlice(block_id=10, src_offset_bytes=0x9000, bytes=0x400),
-            ),
-        ),
-        LayerBlockSlices(
-            regions=(
-                BlockRegionSlice(block_id=11, src_offset_bytes=0x1400, bytes=0x400),
-                BlockRegionSlice(block_id=11, src_offset_bytes=0x9400, bytes=0x400),
-            ),
-        ),
-        LayerBlockSlices(
-            regions=(
-                BlockRegionSlice(block_id=13, src_offset_bytes=0x2000, bytes=0x400),
-                BlockRegionSlice(block_id=13, src_offset_bytes=0xA000, bytes=0x400),
-            ),
-        ),
-    ]
-
-    native_blocks = _layer_blocks_to_native(blocks)
-
-    assert native_blocks == [
-        {
-            "regions": [
-                {
-                    "region_idx": 0,
-                    "block_id": 10,
-                    "src_offset_bytes": 0x1000,
-                    "bytes": 0x800,
-                },
-                {
-                    "region_idx": 1,
-                    "block_id": 10,
-                    "src_offset_bytes": 0x9000,
-                    "bytes": 0x800,
-                },
-            ],
-        },
-        {
-            "regions": [
-                {
-                    "region_idx": 0,
-                    "block_id": 13,
-                    "src_offset_bytes": 0x2000,
-                    "bytes": 0x400,
-                },
-                {
-                    "region_idx": 1,
-                    "block_id": 13,
-                    "src_offset_bytes": 0xA000,
-                    "bytes": 0x400,
-                },
-            ],
-        },
-    ]
+# ruff: noqa: F403,F405,I001
+from .pd_connector_test_utils import *
 
 
 def test_pd_worker_wait_handshake_uses_registered_native_mr_desc() -> None:
@@ -969,7 +228,6 @@ def test_pd_worker_publishes_wait_handshake_and_delays_done_until_all_blocks() -
         base_addr=tensor.data_ptr(),
         block_len=4 * 16 * 32 * 2,
     )
-    assert handshake.layers[0].region_block_addrs(0)[0] == (tensor.data_ptr() + 1 * 4 * 16 * 32 * 2)
 
     push_worker = PdWorkerConnector(
         SimpleNamespace(kv_transfer_config=SimpleNamespace(engine_id="")), rdma=MockRdmaPort()
@@ -1278,143 +536,6 @@ def test_d_worker_rank0_dispatches_prefill_on_wait() -> None:
     assert parsed.layers[0].block_ids == (1,)
 
 
-def test_async_prefill_sender_dispatches_requests_in_parallel(monkeypatch) -> None:
-    started: queue.Queue[str] = queue.Queue()
-    finished: queue.Queue[str] = queue.Queue()
-    second_started = threading.Event()
-    release_first = threading.Event()
-
-    def fake_post_prefill_request(task: PrefillHttpTask) -> None:
-        started.put(task.request_id)
-        if task.request_id == "first":
-            assert second_started.wait(timeout=2)
-            assert release_first.wait(timeout=2)
-        else:
-            second_started.set()
-        finished.put(task.request_id)
-
-    monkeypatch.setattr(prefill_mod, "post_prefill_request", fake_post_prefill_request)
-    sender = AsyncPrefillSender(worker_count=2)
-    try:
-        sender.submit(_prefill_http_task("first"))
-        assert started.get(timeout=2) == "first"
-        sender.submit(_prefill_http_task("second"))
-
-        assert second_started.wait(timeout=2)
-        release_first.set()
-        assert {finished.get(timeout=2), finished.get(timeout=2)} == {"first", "second"}
-    finally:
-        release_first.set()
-        sender.close()
-
-
-def test_rdma_done_waiter_dispatches_requests_in_parallel() -> None:
-    class BlockingRdma:
-        def __init__(self) -> None:
-            self.second_started = threading.Event()
-            self.release_first = threading.Event()
-            self.started: queue.Queue[str] = queue.Queue()
-            self.finished: queue.Queue[str] = queue.Queue()
-
-        def wait_done(self, req_id: str) -> None:
-            self.started.put(req_id)
-            if req_id == "first":
-                assert self.second_started.wait(timeout=2)
-                assert self.release_first.wait(timeout=2)
-            else:
-                self.second_started.set()
-            self.finished.put(req_id)
-
-    rdma = BlockingRdma()
-    waiter = decode_worker_mod._AsyncRdmaDoneWaiter(rdma, worker_count=2)
-    try:
-        waiter.submit(
-            decode_worker_mod._RdmaWaitTask(
-                req_id="first",
-                remote_request_id="first-p",
-                done_request_id="first-d",
-                prefill_url="http://p:8001",
-                rank=0,
-                block_count=1,
-                queued_ts_ns=time.time_ns(),
-            )
-        )
-        assert rdma.started.get(timeout=2) == "first"
-        waiter.submit(
-            decode_worker_mod._RdmaWaitTask(
-                req_id="second",
-                remote_request_id="second-p",
-                done_request_id="second-d",
-                prefill_url="http://p:8001",
-                rank=0,
-                block_count=1,
-                queued_ts_ns=time.time_ns(),
-            )
-        )
-
-        assert rdma.second_started.wait(timeout=2)
-        rdma.release_first.set()
-        assert {rdma.finished.get(timeout=2), rdma.finished.get(timeout=2)} == {
-            "first",
-            "second",
-        }
-    finally:
-        rdma.release_first.set()
-        waiter.close()
-
-
-def test_layer_push_sender_does_not_recreate_discarded_stats() -> None:
-    class FailingRdma:
-        def __init__(self) -> None:
-            self.release_slow = threading.Event()
-
-        def push_layer(self, req_id, layer_idx, blocks) -> None:
-            if layer_idx == 0:
-                raise RuntimeError("boom")
-            assert self.release_slow.wait(timeout=2)
-
-    rdma = FailingRdma()
-    sender = prefill_worker_mod._AsyncLayerPushSender()
-    try:
-        sender.submit(
-            prefill_worker_mod._LayerPushTask(
-                rdma=rdma,
-                req_id="req",
-                target_request_id="decode",
-                chunk_idx=1,
-                layer_idx=0,
-                block_count=1,
-                rdma_bytes=1,
-                block_slices=[],
-                enqueued_ts_ns=0,
-            )
-        )
-        sender.submit(
-            prefill_worker_mod._LayerPushTask(
-                rdma=rdma,
-                req_id="req",
-                target_request_id="decode",
-                chunk_idx=1,
-                layer_idx=1,
-                block_count=1,
-                rdma_bytes=1,
-                block_slices=[],
-                enqueued_ts_ns=0,
-            )
-        )
-
-        with pytest.raises(RuntimeError, match="boom"):
-            sender.wait_req("req")
-        sender.discard_req_stats("req")
-        rdma.release_slow.set()
-        sender.wait_all()
-
-        assert sender.pop_req_stats("req").tasks == 0
-    finally:
-        rdma.release_slow.set()
-        sender.close()
-
-
 def test_layer_push_sender_keeps_fifo_order() -> None:
     class Event:
         def __init__(self) -> None:
@@ -1440,13 +561,8 @@ def test_layer_push_sender_keeps_fifo_order() -> None:
             prefill_worker_mod._LayerPushTask(
                 rdma=rdma,
                 req_id="req",
-                target_request_id="decode",
-                chunk_idx=1,
                 layer_idx=0,
-                block_count=1,
-                rdma_bytes=1,
                 block_slices=[],
-                enqueued_ts_ns=0,
                 event=first_event,
             )
         )
@@ -1454,13 +570,8 @@ def test_layer_push_sender_keeps_fifo_order() -> None:
             prefill_worker_mod._LayerPushTask(
                 rdma=rdma,
                 req_id="req",
-                target_request_id="decode",
-                chunk_idx=1,
                 layer_idx=1,
-                block_count=1,
-                rdma_bytes=1,
                 block_slices=[],
-                enqueued_ts_ns=0,
                 event=ready_event,
             )
         )
@@ -1475,40 +586,6 @@ def test_layer_push_sender_keeps_fifo_order() -> None:
     finally:
         first_event.ready.set()
         sender.close()
-
-
-def test_req_push_stats_records_latency_distribution() -> None:
-    stats = prefill_worker_mod._ReqPushStats()
-    for queue_ms, event_ms, native_ms, event_ready_ts_ns, native_done_ts_ns in (
-        (1.0, 10.0, 100.0, 1_000_000_000, 1_100_000_000),
-        (2.0, 20.0, 200.0, 2_000_000_000, 2_200_000_000),
-        (3.0, 30.0, 300.0, 3_000_000_000, 3_300_000_000),
-    ):
-        stats.add(
-            prefill_worker_mod._LayerPushResult(
-                bytes=4,
-                queue_wait_ms=queue_ms,
-                event_ms=event_ms,
-                native_ms=native_ms,
-                event_ready_ts_ns=event_ready_ts_ns,
-                native_start_ts_ns=event_ready_ts_ns,
-                native_done_ts_ns=native_done_ts_ns,
-            )
-        )
-
-    assert stats.tasks == 3
-    assert stats.bytes == 12
-    assert stats.avg_queue_wait_ms() == 2.0
-    assert stats.p50_queue_wait_ms() == 2.0
-    assert stats.p95_queue_wait_ms() == 3.0
-    assert stats.avg_event_ms() == 20.0
-    assert stats.p50_event_ms() == 20.0
-    assert stats.p95_event_ms() == 30.0
-    assert stats.avg_native_ms() == 200.0
-    assert stats.p50_native_ms() == 200.0
-    assert stats.p95_native_ms() == 300.0
-    assert stats.event_ready_window_ms() == 2000.0
-    assert stats.native_submit_window_ms() == 2300.0
 
 
 def _prefill_http_task(request_id: str) -> PrefillHttpTask:
